+++ src/core/entities/entity.go
@@ -4,10 +4,8 @@ import (
 	"github.com/google/uuid"
 )
 
-//ID entity ID
 type ID = uuid.UUID
 
-//NewID create a new entity ID
 func NewID() ID {
 	return ID(uuid.New())
 }