+++ src/core/entities/user/error.go
@@ -2,8 +2,6 @@ package user
 
 import "errors"
 
-//Password incorrect
 var ErrIncorrectPassword = errors.New("password incorrect")
 
-//Not found user error
 var ErrUserNotFound = errors.New("user not found")