+++ src/core/ports/jsonplaceholder-integration.go
@@ -1,0 +1,11 @@
+package ports
+
+import (
+	integration "go-api/src/infra/http/integrations/jsonplaceholder/responses"
+)
+
+type (
+	JsonPlaceholderIntegration interface {
+		GetUsers() ([]integration.User, error)
+	}
+)