+++ src/core/useCases/user/get-user/interface.go
@@ -1,0 +1,11 @@
+package get_user
+
+import (
+	entity "go-api/src/core/entities/user"
+)
+
+type (
+	UseCase interface {
+		GetUser(id string) (*entity.User, error)
+	}
+)