+++ src/core/useCases/user/get-user/service.go
@@ -1,0 +1,17 @@
+package get_user
+
+import (
+	interfaces "go-api/src/core/ports"
+)
+
+type Service struct {
+	RepositoryUser             interfaces.Repository
+	IntegrationJsonPlaceHolder interfaces.JsonPlaceholderIntegration
+}
+
+func NewService(repository interfaces.Repository, jsonPlaceholderIntegration interfaces.JsonPlaceholderIntegration) UseCase {
+	return &Service{
+		RepositoryUser:             repository,
+		IntegrationJsonPlaceHolder: jsonPlaceholderIntegration,
+	}
+}