+++ src/infra/http/integrations/jsonplaceholder/responses/company.go
@@ -1,0 +1,7 @@
+package response_jsonplaceholder
+
+type Company struct {
+	Name        string `json:"name"`
+	CatchPhrase string `json:"catchPhrase"`
+	Bs          string `json:"bs"`
+}