+++ src/main/main.go
@@ -12,16 +12,16 @@ func StartApp() {
 		return
 	}
 
-	err = server.ConnectToDabase()
+	err = server.db.ConnectToDabase()
 
 	if err != nil {
 		log.Fatal("failed to create to the database", err)
 		return
 	}
 
-	defer server.CloseDB()
+	defer server.db.CloseDB()
 
-	err = server.Setup().Start()
+	err = server.Setup().engine.Start()
 
 	if err != nil {
 		log.Fatal("failed to serve the api", err)